+++ src/agent/messaging/routing/messages.ts
@@ -1,9 +1,6 @@
+import { MessageType } from '../messagetype';
 import uuid from 'uuid/v4';
 
-export enum MessageType {
-  RouteUpdateMessage = 'did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/routecoordination/1.0/keylist_update',
-  ForwardMessage = 'did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/routing/1.0/forward',
-}
 
 export function createRouteUpdateMessage(recipientKey: Verkey) {
   return {
